// packages/libs/api/src/utils/mongo/ErrorCodes.ts
export enum ErrorCodes {
  NOT_FOUND = 'not-found',
  INTERNAL_FIELDS_IN_BODY = 'internal-fields',
  NOT_WELL_DEFINED = 'not-well-defined',
  COULD_NOT_INSERT = 'could-not-insert',
  COULD_NOT_UPDATE = 'could-not-update',
  RUNTIME_ERROR = 'runtime-error',
  NOT_INITIALIZED = 'not-initialized',
  NOT_IMPLEMENTED = 'not-implemented'
}